+++ source/core/scene/light/null_light.hpp
@@ -31,7 +31,7 @@ class Null_light : public Light {
                 uint32_t sampler_dimension, math::AABB const& bounds, Worker const& worker,
                 Sample_from& result) const noexcept override final;
 
-    float3 evaluate(Sample_from const& sample, float time, Sampler_filter filter,
+    float3 evaluate(Sample_from const& sample, uint64_t time, Sampler_filter filter,
                     Worker const& worker) const noexcept override final;
 
     float pdf(Ray const& ray, Intersection const& intersection, bool total_sphere,
