+++ source/core/scene/material/display/display_constant.cpp
@@ -32,7 +32,7 @@ material::Sample const& Constant::sample(float3 const& wo, Renderstate const& rs
 }
 
 float3 Constant::evaluate_radiance(float3 const& /*wi*/, float2 /*uv*/, float /*area*/,
-                                   float /*time*/, Sampler_filter /*filter*/,
+                                   uint64_t /*time*/, Sampler_filter /*filter*/,
                                    Worker const& /*worker*/) const noexcept {
     return emission_;
 }
