+++ source/core/scene/material/light/light_constant.cpp
@@ -25,7 +25,7 @@ material::Sample const& Constant::sample(float3 const& wo, Renderstate const& rs
     return sample;
 }
 
-float3 Constant::evaluate_radiance(float3 const& /*wi*/, float2 /*uv*/, float area, float /*time*/,
+float3 Constant::evaluate_radiance(float3 const& /*wi*/, float2 /*uv*/, float area, uint64_t /*time*/,
                                    Sampler_filter /*filter*/, Worker const& /*worker*/) const
     noexcept {
     return emittance_.radiance(area);
