+++ source/core/scene/material/light/light_emissionmap.cpp
@@ -37,7 +37,7 @@ material::Sample const& Emissionmap::sample(float3 const& wo, Renderstate const&
 }
 
 float3 Emissionmap::evaluate_radiance(float3 const& /*wi*/, float2   uv, float /*area*/,
-                                      float /*time*/, Sampler_filter filter,
+                                      uint64_t /*time*/, Sampler_filter filter,
                                       Worker const& worker) const noexcept {
     auto& sampler = worker.sampler_2D(sampler_key(), filter);
