+++ source/core/scene/material/sky/sky_material_overcast.hpp
@@ -14,7 +14,7 @@ class Material_overcast : public Material {
                                    sampler::Sampler& sampler, Worker const& worker,
                                    uint32_t depth) const noexcept override final;
 
-    float3 evaluate_radiance(float3 const& wi, float2 uv, float area, float time,
+    float3 evaluate_radiance(float3 const& wi, float2 uv, float area, uint64_t time,
                              Sampler_filter filter, Worker const& worker) const
         noexcept override final;
