+++ source/core/scene/material/substitute/substitute_base_material.cpp
@@ -13,7 +13,7 @@ Material_base::Material_base(Sampler_settings const& sampler_settings, bool two_
     : material::Material(sampler_settings, two_sided) {}
 
 float3 Material_base::evaluate_radiance(float3 const& /*wi*/, float2   uv, float /*area*/,
-                                        float /*time*/, Sampler_filter filter,
+                                        uint64_t /*time*/, Sampler_filter filter,
                                         Worker const& worker) const noexcept {
     if (emission_map_.is_valid()) {
         // For some reason Clang needs this to find inherited Material::sampler_key_
