+++ source/core/scene/material/substitute/substitute_base_material.hpp
@@ -9,7 +9,7 @@ class Material_base : public material::Material {
   public:
     Material_base(Sampler_settings const& sampler_settings, bool two_sided) noexcept;
 
-    float3 evaluate_radiance(float3 const& wi, float2 uv, float area, float time,
+    float3 evaluate_radiance(float3 const& wi, float2 uv, float area, uint64_t time,
                              Sampler_filter filter, Worker const& worker) const
         noexcept override final;
