+++ source/extension/procedural/sky/sky_material.cpp
@@ -43,7 +43,7 @@ material::Sample const& Sky_material::sample(float3 const& wo, Renderstate const
 }
 
 float3 Sky_material::evaluate_radiance(float3 const& wi, float2 /*uv*/, float /*area*/,
-                                       float /*time*/, Sampler_filter /*filter*/,
+                                       uint64_t /*time*/, Sampler_filter /*filter*/,
                                        Worker const& /*worker*/) const noexcept {
     return sky_.model().evaluate_sky(wi);
 }
@@ -86,7 +86,7 @@ material::Sample const& Sky_baked_material::sample(float3 const& wo, Renderstate
 }
 
 float3 Sky_baked_material::evaluate_radiance(float3 const& /*wi*/, float2   uv, float /*area*/,
-                                             float /*time*/, Sampler_filter filter,
+                                             uint64_t /*time*/, Sampler_filter filter,
                                              Worker const& worker) const noexcept {
     auto const& sampler = worker.sampler_2D(sampler_key(), filter);
     return emission_map_.sample_3(sampler, uv);
